Learnify/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.1.1
django-allauth==64.2.1
django-braces==1.15.0
django-debug-toolbar==4.4.6
django-embed-video==1.4.10
djangorestframework==3.15.2
djlint==1.35.2
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.8
jsbeautifier==1.15.1
json5==0.9.25
pathspec==0.12.1
pillow==10.4.0
psycopg2==2.9.9
pymemcache==4.0.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
